3단계 - 자동차 경주 : view, domain 분리
