+++ src/main/kotlin/study/step3/domain/Car.kt
@@ -1,4 +1,4 @@
-package study.step3
+package study.step3.domain
 
 class Car {
