+++ src/main/kotlin/study/step3/view/InputView.kt
@@ -1,4 +1,4 @@
-package study.step3
+package study.step3.view
 
 object InputView {
