+++ src/main/kotlin/study/step3/view/ResultView.kt
@@ -1,4 +1,6 @@
-package study.step3
+package study.step3.view
+
+import study.step3.domain.RacingCar
 
 object ResultView {
     fun printResult() {
